+++ tests/CmsConsoleTestCase.php
@@ -24,7 +24,10 @@ class CmsConsoleTestCase extends ConsoleApplicationTestCase
                 ]
             ],
             'modules' => [
-                'admin' => 'luya\admin\Module',
+                'admin' => [
+                    'class' => 'luya\admin\Module',
+                    'queueMutexClass' => 'yii\mutex\FileMutex',
+                ],
                 'cms' => [
                     'class' => '\luya\cms\frontend\Module'
                 ],