+++ CHANGELOG.md
@@ -6,6 +6,11 @@ In order to read more about upgrading and BC breaks have a look at the [UPGRADE 
 ## 4.0.0
 
 + [#320](https://github.com/luyadev/luya-module-cms/pull/320) Replace cms menu item `publish_from` and `publish_till` with sheduler.
++ [#329](https://github.com/luyadev/luya-module-cms/issues/329) Ensure its not possible to drag a placeholder block into itself (this can create circular references)
++ [#325](https://github.com/luyadev/luya-module-cms/issues/325) Added target value into the NavTree widget.
++ [#327](https://github.com/luyadev/luya-module-cms/issues/327) Fix LinkConverter class when using CMS as headless API.
++ [#323](https://github.com/luyadev/luya-module-cms/pull/323) Add LUYA Test Suite 2.0 with PHP 8 Support.
++ [#321](https://github.com/luyadev/luya-module-cms/pull/321) Added scheduler for block item visibility.
 
 ## 3.5.1 (24. November 2020)